Introduce monitor.rs package, factor out helpers there
