+++ crates/node/src/indexer.rs
@@ -35,6 +35,7 @@ pub mod configs;
 pub mod foreign_chain;
 pub mod handler;
 pub mod migrations;
+mod monitor;
 pub mod near_data_wipe;
 pub mod participants;
 pub mod real;
